fix(client): guard hex-colour parsing and unify file-size formatting
