+++ apps/client/lib/src/widgets/chat_input_bar/file_pickers.dart
@@ -6,6 +6,7 @@ import 'package:flutter/foundation.dart' show debugPrint, kIsWeb;
 import 'package:flutter/material.dart';
 
 import '../../services/toast_service.dart';
+import '../../utils/byte_format.dart';
 import 'upload_helpers.dart';
 
 typedef StageAttachmentFn =
